MeLOne.accuracy_score — Method

accuracy_score(y_real, y_pred)

Returns the amount of times y_pred is equal to y_real as a percentage.

MeLOne.fit! — Function

fit!(model, X, y)

Fit the model with matrix X and target y.

MeLOne.predict — Function

y_pred = predict(model, X)

Predict the target of matrix X with model. y_pred is a Vector.

MeLOne.predict_proba — Function

prob_mat = predict_proba(model, X)

Predict the probability of each row of matrix X being of each class in model._classes. prob_mat is a matrix with dimensions size(X, 1) by length(model._classes). Each row of prob_mat that sum 1.

MeLOne.r2_score — Method

r2_score(y_real, y_pred)

Computes the R² score given by $1 - ∑(yᵢ - ŷᵢ)² / ∑(yᵢ - ȳ)²$, where y is y_real, ŷ is y_pred and ȳ is the mean of y_real.

MeLOne.rss_score — Method

rss_score(y_real, y_pred)

Computes the residual sum of squares.

MeLOne.gini — Method

IG = 1 - ∑pₖ²

MeLOne.gini_split — Method

Escolher qual coluna e threshold para cortar o conjunto I em dois.

MeLOne.regressor_impurity — Method

regressor_impurity(y)

∑ᵢ(yᵢ - ȳ)² / n

MeLOne.regressor_split — Method

Escolher qual coluna e threshold para cortar o conjunto I em dois.